Conflicts (resolved):
	client/X11/xf_gdi.c
	libfreerdp/gdi/16bpp.c
	libfreerdp/gdi/32bpp.c
	libfreerdp/gdi/8bpp.c

+++ winpr/libwinpr/sspi/test/TestEnumerateSecurityPackages.c
@@ -23,7 +23,7 @@ int TestEnumerateSecurityPackages(int argc, char* argv[])
 
 	printf("\nEnumerateSecurityPackages (%d):\n", (unsigned int)cPackages);
 
-	for (index = 0; index < cPackages; index++)
+	for (index = 0; index < (int) cPackages; index++)
 	{
 		printf("\"%s\", \"%s\"\n", pPackageInfo[index].Name, pPackageInfo[index].Comment);
 	}
